+++ src/Domain/Automation/Support/Livewire/Actions/WaitActionComponent.php
@@ -2,14 +2,15 @@
 
 namespace Spatie\Mailcoach\Domain\Automation\Support\Livewire\Actions;
 
+use Carbon\CarbonInterval;
 use Illuminate\Validation\Rule;
 use Spatie\Mailcoach\Domain\Automation\Support\Livewire\AutomationActionComponent;
 
 class WaitActionComponent extends AutomationActionComponent
 {
-    public string $length = '1';
+    public ?string $length = '1';
 
-    public string $unit = 'days';
+    public ?string $unit = 'days';
 
     public array $units = [
         'minutes' => 'Minute',
@@ -19,11 +20,21 @@ class WaitActionComponent extends AutomationActionComponent
         'months' => 'Month',
     ];
 
+    public function mount()
+    {
+        $this->length ??= '1';
+        $this->unit ??= 'days';
+    }
+
     public function getData(): array
     {
+        $unit = $this->unit;
+        $interval = CarbonInterval::$unit($this->length);
+
         return [
-            'length' => (int) $this->length,
+            'seconds' => $interval->totalSeconds,
             'unit' => $this->unit,
+            'length' => $this->length,
         ];
     }