+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `laravel-mailcoach` will be documented in this file
 ## 4.8.5 - 2021-08-16
 
 - Fix an issue with pagination on queries when `created_at` has the same value. (#671)
+- Fix the WaitAction throwing errors with Carbon localization (#657)
 
 ## 4.8.4 - 2021-08-06